crawler/pyproject.toml
[tool.poetry]
name = "crawler"
version = "0.1.0"
description = ""
authors = ["jiho7407 <mmm7407@yonsei.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
